+++ arcore-location-extension/src/main/kotlin/org/neidhardt/arlocation/LocationArNode.kt
@@ -84,7 +84,7 @@ class LocationArNode(
 					markerLocation
 			).ellipsoidalDistance
 
-			val distanceScaleFactor = ArUtils.scaleFactorForDistance(distance)
+			val distanceScaleFactor = ArUtils.linearScaleFactor(distance)
 			return scaleFixedSize * distanceScaleFactor
 		}
 	}