+++ arcore-location-extension/src/test/kotlin/org/neidhardt/arlocation/LocationArNodeTest.kt
@@ -34,7 +34,7 @@ class LocationArNodeTest {
 		)
 		// verify
 		assertEquals(
-				LocationArNode.getScaleFixedSize(testDirection) * ArUtils.scaleFactorForDistance(distance),
+				LocationArNode.getScaleFixedSize(testDirection) * ArUtils.linearScaleFactor(distance),
 				result,
 				0.1
 		)